Created socket routes for typing adn messages
- When user writes a message, send button pushes
it through socket to server. Server socket then
emits message and name of the sender to all
sockets
- When user has focus on input message field,
it triggers socket route typing with data name.
Server broadcasts
trigger to all OTHER socket connections that
someone is writing.
- When user unfocuses input message window,
route is triggered that brodcasts to all OTHER
socket connections that the original socket is
no longer typing